qol(Solitaire)
Main: neatened the menu and display structure for the solitaire gamemode, and added basic functionality to pick up and move cards, but without checks for valid moves yet.

Card: added function to read card colour, and function to compare how similar two cards are

Deck: added function to reverse the order of the cards in a deck, and added a display option for vertical spread with a smaller gap.

UI: added function to read position of menu cursor

added light/dark grey colour values, and fixed some existing colours